engine: make discovery retention exact and cover stockless models
Five review findings on the discovery-exact enumeration pipeline, all
touching ltm_finding.rs/ltm_finding_enum.rs's retention and universe
accounting:

- A model with no parent-level stocks no longer short-circuits to an
  empty universe. The enumerator needs no stock seeds at all, and the
  fallback's default seed policy already seeds stockless SCCs
  directly, so a 2+-node cycle whose only state is a module level or a
  PREVIOUS lag between auxes (a real feedback loop) is now analyzed
  and reported in a Solo group under both generators, instead of being
  declared empty by construction.

- Retention now decides, before either candidate's mass reaches a
  partition total, which enumerated circuits are non-representative
  twins of another circuit that trims to the identical reported loop
  (a direct reference and its hoisted-reducer twin, AC4.3). Previously
  this was corrected only after materialization, which meant
  retention's own MIN_CONTRIBUTION threshold judged every circuit
  against an inflated total and could drop a genuinely-reportable loop
  that would have cleared the corrected denominator. The new
  dedup_trimmed_twins pass groups circuits by their trimmed node
  identity (agg-bearing circuits directly; non-agg circuits are tested
  against that small key set rather than materialized wholesale, so a
  model with no synthetic agg node -- most of the corpus -- pays
  nothing) and keeps the higher-raw_avg_abs_score representative,
  matching the tie semantics the post-materialization dedup already
  used. The post-materialization dedup and its totals correction stay
  as a safety net for the two cases retention still cannot see: a
  module-traversing twin (whose reported score is the per-exit-port
  override, not this raw product) and a stitched cross-agg loop.

- subtract_reported_mass_from_totals no longer computes Inf - Inf: at
  a dominance inflection the running total is already Inf by
  convention, and if the dropped duplicate's own score is also Inf
  there, subtracting it would poison the total with NaN for every
  sibling loop at that step. The subtraction is now skipped at such a
  step, leaving the total Inf.

- Retention's universe-circuit count no longer includes a circuit
  whose individually-active edges multiply to a product that
  underflows (or Inf*0-NaNs) to exactly zero at every step. Such a
  circuit banks no mass anywhere, so counting it toward
  partition_circuit_counts wrongly inflated a partition to "competing"
  even though only one loop's mass was actually in the denominator.

- RetentionOutcome now carries distinct_circuits (enumerated circuits
  minus retention-time-deduped twins), and DiscoveryResult's
  universe_loops is now that count plus the deduped stitched cross-agg
  loop count -- the number of DISTINCT loops whose mass the partition
  denominators actually sum, rather than the raw enumerated circuit
  count (which never included stitched loops and could overcount
  duplicate twins).

World3 and C-LEARN, which carry no dedup twins, are unaffected in
survivor/universe counts and stay within the AC3.1 time budget
(0.436s / 0.039s in release, well under 1.0s / 0.2s).

ltm_finding_tests.rs split into a second sibling `#[path]` module
(ltm_finding_enum_tests.rs, the union-graph enumeration/retention
suite) to stay under the 6000-line per-file lint cap after these
additions.
